refactor(domain): corrige as entidades da Sprint 1 e adiciona usuários e perfis
As entidades só tinham getters, mas os controllers chamavam setters, e os repositories usavam findAllByAtivoTrue em entidades sem esse campo: o projeto não compilava. As entidades ganham comportamento (substituirDados/PUT, atualizarInformacoes/PATCH, excluir), tipos alinhados ao schema (LocalDate, BigDecimal, enums) e validações (@CPF, @CNPJ, VIN ISO 3779). Novos: Usuario (UserDetails) com perfis ADMIN, ANALISTA e CONCESSIONARIA, RefreshToken e exceções de domínio.

+++ src/main/java/br/com/ford/vinshare/domain/veiculo/DadosDetalhamentoVeiculo.java
@@ -1,5 +1,7 @@
 package br.com.ford.vinshare.domain.veiculo;
 
+import java.math.BigDecimal;
+
 public record DadosDetalhamentoVeiculo(
         Long id,
         String vin,
@@ -9,7 +11,7 @@ public record DadosDetalhamentoVeiculo(
         Integer anoModelo,
         String cor,
         String combustivel,
-        Double valorCompra,
+        BigDecimal valorCompra,
         String tipoVeiculo
 ) {
     public DadosDetalhamentoVeiculo(Veiculo veiculo) {
